requirements.txt
ruamel.yaml==0.17.4
torch==1.9.0
torchvision==0.9.1
transformers==4.5.1
datasets==1.6.2
tensorboard==2.5.0
nlcodec==0.4.0
ray[tune,default]==1.4.1